refactor(storage): reorganize storage strategies into dedicated module with factory pattern
