+++ docs/demos/tools/conversation/IndexedDB.vue
@@ -23,7 +23,7 @@
 
 <script setup lang="ts">
 import { TrBubbleList, TrSender, BubbleRoleConfig } from '@opentiny/tiny-robot'
-import { useConversation, AIClient, GeneratingStatus } from '@opentiny/tiny-robot-kit'
+import { useConversation, AIClient, GeneratingStatus, indexedDBStorageStrategyFactory } from '@opentiny/tiny-robot-kit'
 import { IconAi, IconUser } from '@opentiny/tiny-robot-svgs'
 import { TinySelect, TinyButton } from '@opentiny/vue'
 import { computed, h } from 'vue'
@@ -58,11 +58,10 @@ const {
   switchConversation,
 } = useConversation({
   client,
-  storageType: 'indexedDB',
-  storageConfig: {
+  storage: indexedDBStorageStrategyFactory({
     dbName: 'demo-chat-db', // 自定义数据库名称
     dbVersion: 1, // 数据库版本
-  },
+  }),
   events: {
     onLoaded(conversations) {
       if (conversations.length === 0) {
