+++ packages/kit/src/storage/index.ts
@@ -1,0 +1,9 @@
+export { LocalStorageStrategy } from './LocalStorageStrategy'
+export { IndexedDBStrategy } from './IndexedDBStrategy'
+export {
+  localStorageStrategyFactory,
+  indexedDBStorageStrategyFactory,
+  type LocalStorageConfig,
+  type IndexedDBConfig,
+} from './factories'
+export type { ConversationStorageStrategy } from './types'
