--- packages/kit/src/vue/conversation/storage/index.ts
@@ -1,4 +1,0 @@
-export { LocalStorageStrategy } from './LocalStorageStrategy'
-export { IndexedDBStrategy } from './IndexedDBStrategy'
-export { createStorageStrategy } from './createStrategy'
-export type { ConversationStorageStrategy, StorageConfig, StorageType } from './types'
