+++ packages/kit/src/vue/index.ts
@@ -1,4 +1,11 @@
 export * from './message/useMessage'
 export * from './conversation/useConversation'
-export { LocalStorageStrategy, IndexedDBStrategy, createStorageStrategy } from './conversation/storage'
-export type { ConversationStorageStrategy, StorageConfig, StorageType } from './conversation/storage'
+export {
+  LocalStorageStrategy,
+  IndexedDBStrategy,
+  localStorageStrategyFactory,
+  indexedDBStorageStrategyFactory,
+  type LocalStorageConfig,
+  type IndexedDBConfig,
+  type ConversationStorageStrategy,
+} from '../storage'
